requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
requests
streamlit
streamlit-autorefresh
matplotlib
openai
python-dotenv
sqlalchemy
seaborn
huggingface-hub